+++ src/RadialGauge/RadialGaugeSeries/RadialGaugeLabel.tsx
@@ -1,4 +1,4 @@
-import React, { PureComponent } from 'react';
+import React, { FC } from 'react';
 import {
   ChartShallowDataShape,
   ChartInternalDataTypes
@@ -29,27 +29,25 @@ export interface RadialGaugeLabelProps {
   onClick: (event: { data; nativeEvent }) => void;
 }
 
-export class RadialGaugeLabel extends PureComponent<RadialGaugeLabelProps> {
-  static defaultProps: Partial<RadialGaugeLabelProps> = {
-    onClick: () => undefined
-  };
+export const RadialGaugeLabel: FC<Partial<RadialGaugeLabelProps>> = ({
+  data,
+  className,
+  offset,
+  onClick = () => undefined
+}) => {
+  const label = formatValue(data.key as ChartInternalDataTypes);
 
-  render() {
-    const { data, className, offset, onClick } = this.props;
-    const label = formatValue(data.key as ChartInternalDataTypes);
-
-    return (
-      <text
-        x="0"
-        y={offset}
-        textAnchor="middle"
-        // This is only valid when placed below the chart
-        alignmentBaseline={'text-after-edge'}
-        onClick={(nativeEvent) => onClick({ data, nativeEvent })}
-        className={classNames(className, css.valueLabel)}
-      >
-        {label}
-      </text>
-    );
-  }
-}
+  return (
+    <text
+      x="0"
+      y={offset}
+      textAnchor="middle"
+      // This is only valid when placed below the chart
+      alignmentBaseline={'text-after-edge'}
+      onClick={(nativeEvent) => onClick({ data, nativeEvent })}
+      className={classNames(className, css.valueLabel)}
+    >
+      {label}
+    </text>
+  );
+};